- (Re-)include residential satisfaction in dwelling-search - Corrected bug in unique filename creation for output-files

+++ src/at/sume/dm/model/residential_mobility/ResidentialMobility.java
@@ -82,7 +82,7 @@ public class ResidentialMobility {
 	public DwellingRow searchDwelling(HouseholdRow household, int modelYear, DwellingsOnMarket dwellingsOnMarket, boolean considerDwellingCosts) {
 		// Adjust the search area size to the number of years the household has been looking for a new dwelling
 		int searchAreaSize = Common.getSearchAreaSize() + (modelYear - household.getMovingDecisionYear()) * Common.getSearchAreaSizeIncement();
-		ArrayList<Long> potentialTargetSpatialUnitIds = household.getPreferredSpatialUnits(searchAreaSize);
+		ArrayList<Integer> potentialTargetSpatialUnitIds = household.getPreferredSpatialUnits(searchAreaSize);
 		assert potentialTargetSpatialUnitIds.size() > 0 : "no potential target spatial units found";
 		int suitableDwellingCount = 0;
 		if (considerDwellingCosts) {
